% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{df.test}
\alias{df.test}
\title{Augmented Dickey_Fuller test}
\usage{
df.test(ts, lag)
}
\arguments{
\item{ts}{Numerical dataframe.}

\item{lag}{The lag parameter.}
}
\value{
df: returns the value of the test.

summary ():  shows the test results.
}
\description{
Augmented Dickey_Fuller test
}
\details{
Computes the stationarity test for a given univariate time series.
}
\examples{
library (timeSeries)
library (NlinTS)
#load data
data = LPP2005REC
model = df.test (data[,1], 1)
model$summary ()
}
\references{
{
  \insertRef{elliott1992efficient}{NlinTS}
}
}
