% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{entropy_cont}
\alias{entropy_cont}
\title{Continuous  entropy}
\usage{
entropy_cont(V, k = 3, log = "loge")
}
\arguments{
\item{V}{Interger vector.}

\item{k}{Integer argument specifying the number of neighbors.}

\item{log}{A string specifying the logarithm base: "log2", "loge" (default), or "log10".}
}
\description{
Continuous  entropy
}
\details{
Computes the continuous entropy of a numerical vector using the Kozachenko approximation.
}
\examples{
library (timeSeries)
library (NlinTS)
#load data
data = LPP2005REC
print (entropy_cont (data[,1], 3))
}
\references{
{
  \insertRef{kraskov2004estimating}{NlinTS}
}
}
