% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{entropy_disc}
\alias{entropy_disc}
\title{Discrete Entropy}
\usage{
entropy_disc(V, log = "log2")
}
\arguments{
\item{V}{Integer vector.}

\item{log}{A string specifying the logarithm base for entropy calculation: "log2" (default), "loge", or "log10".}
}
\description{
Discrete Entropy
}
\details{
Computes the Shanon entropy of an integer vector.
}
\examples{
library (NlinTS)
print (entropy_disc (c(3,2,4,4,3)))
}
