% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{entropy_cont}
\alias{entropy_cont}
\title{Continuous  entropy}
\usage{
entropy_cont(V, k = 3, log = "loge")
}
\arguments{
\item{V}{Interger vector.}

\item{k}{Integer argument, the number of neighbors.}

\item{log}{String argument in the set ("log2", "loge","log10"), which indicates the log function to use. The loge is used by default.}
}
\description{
Continuous  entropy
}
\details{
Computes the continuous entropy of a numerical vector using the Kozachenko approximation.
}
\examples{
library (timeSeries)
library (NlinTS)
#load data
data = LPP2005REC
print (entropy_cont (data[,1], 3))
}
\references{
{
  \insertRef{kraskov2004estimating}{NlinTS}
}
}
