% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modues.R
\name{entropy_cont}
\alias{entropy_cont}
\title{Continuous  entropy}
\usage{
entropy_cont(V, k = 3)
}
\arguments{
\item{V}{Interger vector.}

\item{k}{Integer argument, the number of neighbours.}
}
\description{
Continuous  entropy
}
\details{
Computes the continuous entropy of a numerical vector using the Kozachenko approximation.
}
\examples{
library (timeSeries)
library (NlinTS)
#load data
data = LPP2005REC
print (entropy_cont (data[,1], 3))
}
\references{
{
  \insertRef{kraskov2004estimating}{NlinTS}
}
}
