% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{entropy_disc}
\alias{entropy_disc}
\title{Discrete Entropy}
\usage{
entropy_disc(V, log = "log2")
}
\arguments{
\item{V}{Integer vector.}

\item{log}{String argument in the set ("log2", "loge","log10"), which indicates the log function to use. The log2 is used by default.}
}
\description{
Discrete Entropy
}
\details{
Computes the Shanon entropy of an integer vector.
}
\examples{
library (NlinTS)
print (entropy_disc (c(3,2,4,4,3)))
}
