% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mi_disc}
\alias{mi_disc}
\title{Discrete multivariate Mutual Information}
\usage{
mi_disc(df, log = "log2", normalize = FALSE)
}
\arguments{
\item{df}{Datafame of type Integer.}

\item{log}{String argument in the set ("log2", "loge","log10"), which indicates the log function to use. The log2 is used by default.}

\item{normalize}{Logical argument (FALSE by default)  for the option of normalizing the mutual information by dividing it by the joint entropy.}
}
\description{
Discrete multivariate Mutual Information
}
\details{
Computes the Mutual Information between columns of a dataframe.
}
\examples{
library (NlinTS)
df = data.frame (c(3,2,4,4,3), c(1,4,4,3,3))
mi = mi_disc (df)
print (mi)
}
