% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mi_disc_bi}
\alias{mi_disc_bi}
\title{Discrete bivariate Mutual Information}
\usage{
mi_disc_bi(X, Y, log = "log2", normalize = FALSE)
}
\arguments{
\item{X}{Integer vector.}

\item{Y}{Integer vector.}

\item{log}{String argument in the set ("log2", "loge","log10"), which indicates the log function to use. The log2 is used by default.}

\item{normalize}{Logical argument (FALSE by default)  for the option of normalizing the mutual information by dividing it by the joint entropy.}
}
\description{
Discrete bivariate Mutual Information
}
\details{
Computes the Mutual Information between two integer vectors.
}
\examples{
library (NlinTS)
mi = mi_disc_bi (c(3,2,4,4,3), c(1,4,4,3,3))
print (mi)
}
