\name{NlinTS-package}
\alias{NlinTS-package}
\alias{NlinTS}
\docType{package}
\title{
Models for non-linear causality detection in time series.
}
\description{
Globally, this package focuses on non-linear time series analysis,  especially on causality detection.
To deal with non-linear dependencies between time series, we propose an extension of the Granger causality test using feed-forward neural networks. This package includes also an implementation of the Transfer Entropy, which  can be also seen as a causality measure based on information theory. To do that, the package  includes discrete and continuous Transfer entropy using the Kraskov approximation. The NlinTS package includes also some other useful tools, like the VARNN (Vector Auto-Regressive Neural Network) model, the Augmented Dickey-Fuller test of stationarity, and the discrete and continuous entropy and mutual information.
}
